feat(shoppingList): Refactor ShoppingList state management using NgRx
This commit integrates the ShoppingList reducer into the NgRx store in the 'app.module.ts' file, enhancing the state management of the ShoppingList component. The Ingredient interface in 'ingredient.interface.ts' is updated to make the 'name' field mandatory.

Furthermore, the 'Add' action in 'shoppingList.actions.ts' is renamed to 'AddIngredient' and is now created using the 'createAction' function from NgRx, allowing for more convenient action creation.

Lastly, the 'shoppingList.reducer.ts' file is refactored to use 'createReducer' and 'on' functions from NgRx, improving the readability and maintainability of the reducer function.