# Cleaned Up 'auth.effects.ts'
This commit removes all commented-out code in 'auth.effects.ts'. The file now begins with necessary imports. Commented code can clutter the file and make it hard to read, so it's best practice to remove it once it's no longer useful.

The file continues to import several modules and objects, which are necessary for the functioning of the application, such as 'Injectable' from '@angular/core', 'HttpClient' from '@angular/common/http', and 'Actions', 'ofType', 'createEffect' from '@ngrx/effects'.